Put IPC command and event names under one scheme
Commands mixed verb-object with object-verb, and events named their
subject sometimes as `window_`, sometimes as `client_`, and sometimes
not at all.  A command is now its verb, then its object, then any
direction, and an event its subject, then what happened to it, with
`_set`, `_cleared` and `_changed`.

+++ include/enact/internal.h
@@ -47,7 +47,7 @@
  * @note No-op if @p client is @c NULL
  * @note Complexity: @e O(1)
  */
-void enact_broadcast_client_event(client_td *client, uint32_t type);
+void enact_broadcast_client_event(client_td *client, uint64_t type);
 
 
 #endif /* ! ENACT_INTERNAL_H */
